requirements.txt
astroid==2.2.5
autopep8==1.4.4
bcrypt==3.1.6
blinker==1.4
cffi==1.12.3
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.5
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.2.1
